+++ packages/widgetbook/lib/src/knobs/knob.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/widgets.dart';
-import 'package:widgetbook_core/widgetbook_core.dart';
 
 import '../fields/field.dart';
-import '../models/models.dart';
+import '../navigation/navigation.dart';
+import '../settings/settings.dart';
 import '../state/state.dart';
 
 /// Allows [WidgetbookUseCase]s to have dynamically adjustable parameters.