+++ packages/widgetbook/test/src/models/widgetbook_component_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:widgetbook/src/models/models.dart';
+import 'package:widgetbook/src/navigation/navigation.dart';
 
 void main() {
   group('$WidgetbookComponent', () {