+++ widgetbook_for_widgetbook/lib/settings/setting_header.widgetbook.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+import 'package:widgetbook/src/settings/settings.dart';
+import 'package:widgetbook/widgetbook.dart';
+import 'package:widgetbook_annotation/widgetbook_annotation.dart';
+
+@UseCase(name: 'Text only', type: SettingHeader)
+Widget textOnly(BuildContext context) {
+  return SettingHeader(
+    content: context.knobs.string(label: 'Content', initialValue: 'Frames'),
+  );
+}
+
+@UseCase(name: 'with Trailing', type: SettingHeader)
+Widget trailing(BuildContext context) {
+  return SettingHeader(
+    content: context.knobs.string(label: 'Content', initialValue: 'Frames'),
+    trailing: Switch(
+      value: context.knobs.boolean(label: 'Trailing'),
+      onChanged: (_) {},
+    ),
+  );
+}